Updated AMRT to 1.3, removed 1.2
Known Bug: Pressing the Green Flag while loop is enabled will cause the renderer to play nothing. This issue is not seen in the HTML-ified versions, however. This bug is found in AMRT 1.3 and AMP 2.3.0+ (or any AMP version with Loop Functionality.)

Fixed bugs and issues seen in the previous AMRT versions.